Consolidate tests for campaign-display components
