--- elements/campaign-display/campaign-display-schema.js
@@ -1,6 +1,0 @@
-import CampaignField from './fields/campaign-field';
-import CampaignRequest from './fields/campaign-request-field';
-export default {
-  campaign: CampaignField,
-  campaignRequest: CampaignRequest,
-};
